Moved all python into ethsnarks module
