+++ .gitignore
@@ -1,4 +1,5 @@
 build/
+*.pyc
 *.swp
 *.swo
 /Makefile
