+++ AUTHORS
@@ -1,4 +1,4 @@
-# This is the official list of semaphore authors for copyright purposes.
+# This is the official list of EthSnarks authors for copyright purposes.
 
 barrywhitehat <barrywhitehat@protonmail.com>
 HarryR - https://github.com/HarryR
