+++ zksnark_element/.gitignore
@@ -1,1 +1,2 @@
 vk.raw
+pk.json
