+++ zksnark_element/vk.json
@@ -1,14 +1,14 @@
 {
- "a" :[["0x289eddfbcea769b1b9a63af7922fa010d1acab80cf4dfce7946bdb6c0eea2421", "0x1a92b8b58fa253ff54d2387b7737bbcc210bbb1f697812deb9ebd91b48935428"],
- ["0x2064eb0c7ef625adcd74df5a9e95c39c8333bbe72197b2cde295cde053a7492b", "0x17e97c70e4210a5a46ebd57bfe66b01e25a2febe01cc373cc0d5741c5831fa5d"]],
- "b"  :["0x42e25419bc2e4e99ef2b4be64731a27711c161ba81e0119e291692311aa29d0", "0x2aa2b91dcff752cc6697ada42312ff909c04b755e98bbf81953f1e7b18e77996"],
- "c" :[["0x2cebc76b585083c27517c2499e669b53f36a0f1d2c2c660ba066adca19d87851", "0x39e8ae1a7997ee121730df0f2cae75f340551723179334a6e4a87fa3064bf06"],
- ["0x1969ea23d3ef20d3cd5c88a2f80308573c6f32a9abcdc9cc27abf2adeab7e8d0", "0x661d8372bfc17d10ec215cb94ab55773a6c16304e11fa9e8a761c18d7673e80"]],
- "g" :[["0xa8ad7b4a5f81bedccd6a9e9d232da82abedc610a8c3de7fd86a4fbbda7ce4b7", "0x24cbd35b1f9c15aff4b010c5bae95bb638d653055c2a64de60001ab63bcee3df"],
- ["0x181d8408a1fa8f97d0fc7278be3b9b18fc22dd7e3f84cd62611257d6d8b491e8", "0x783f2a5ee5e0191521a58a8bd507e6bf9f37ad9d3b0876bcffe0f48dbc589ab"]],
- "gb1" :["0x5e4263732add8b27c765c07bc2d3b3c4e65c82364907447da8859d2388b9c41", "0x14d8ee3871c78ea472aa3ad1e7c5259de16b264be300f63cba80d5a587d3748"],
- "gb2" :[["0x918c90b22494e32552a51e28a413e57cab7eeac3976b2236a37578f0ff96de1", "0xd49ed175fe273a4aabef82d87d47e463610562763be7bd6b9fb9ddefd2445f5"],
- ["0x14dee545fb36618c3221e22cce4fd9f9940f328b435501a100b489ecb50380ed", "0x10a817c35ef3ad74cc61baf6da44cef8e9916052108bc960176f54d6582e3b6b"]],
- "z" :[["0xfe0eacc0b7b7e4ff3a1b9211043614f1f6d23f7214fb29fe1ba66a0e0f12851", "0x217d36c6fd48929c92d00ef72fe95cbb91342e9d63672d5b3c38325a42e50ba8"],
- ["0x17b602de44943d6168865e671c945bc842320d84425783a0125f5c5bc5083758", "0x1894e77c4278b92c8e4e2b98c051eb291f8e12e5790776486310f2b4063ad946"]],
-"IC" :[["0x27626fee2c20790edfb32e7a503d6ade13c1d31d5b15cb60ea602f836f63a019", "0x2a950e58b607943216bde1e6e1be72a462d1f1a5ae31fc7e5ad0c3f27e490d7b"],["0x101868454a4003da6c28faf14431db185ae58aa77f35e7ff83e1217e4b51bfae", "0x16e362c4d2d5f41674294bebf781eb0a8ecf1e21bbc9bcbdbac8a984d51d3eec"],["0x303664ff8d2c2ea9bd76272caf6d53929a9cdbb8427ef7a179354d113b6f7a57", "0x26d2edc0b24e67922f5fb328c153b33897a61c505f5cef688ed0f5e1937c4110"],["0x7a8d0847370a8f360f6450fe405187c218786ed01a3bd26a78ba35610fc7861", "0x3d3e0d6ca7d2d18d185614f9e75017161a3588098e234f42cf28b8a51392f06"],["0xb9ed4c5fab3c2a2383ca22a7e56e05cfbe80066133c4b518175bf777dc1ec51", "0x1b15e05a1e11362ef472fa90690a4f8b4d55e97e46b9e9e89e35b91db8bcc041"],["0x1bbd55383f24b297ff234df6c151a9497d69b93bcd23733037032eee3969c453", "0x1871d89e0ace668c954d6ad6680d19c6faaa3dc42b24e4fea42c6ed5d3c675e1"]]}
+ "a" :[["0x12502d4998e189cc2e0f32cace5e0fa46133e40f9ce7c17d6009a2356d8b348a", "0x2071b47eba339b74a9c33904b8d3f0ad36f0779ddd4947de9b45fe7b61fb848c"],
+ ["0x1f17142d98b38e614456a58fafdf75f785425e65952117da6c9bf9e5d897f072", "0x278fbb5b40960fb2a8b0b403be98e586383d2641c2311b77cae3e5b1e216c85a"]],
+ "b"  :["66c5f70773bfc62822282349ce0821bc3d17f65bc7d20dc7458328961d7bd29", "0x226481c368d40a432e856522c74e0962e1d8e27915357f67ca97c9f899919fb1"],
+ "c" :[["0x16a5fe78f732ff1ed647425583dad021c05b24f68cea32d546fb58727dc2525a", "0xf2fa8ca7f09173bb249de5986ca38715eaabdffb43d3f6f1307fcc5235499ff"],
+ ["0x2367766772e55794ed9416f700afff4f31274696a1e18d5078d576db0628ef1b", "0xcae3f23e74375736bd4dd4f9c3f936b263650c9e4edee766ca81e1796a18a55"]],
+ "g" :[["0x530f1e1ce2e5d9d7b36221a66dc5dacc742390e88806cbd6b63ff7e1d2119bc", "0x90b30109d5edc605d2d5412040849e23ec92199715c37e66413682c6c921d98"],
+ ["0x20cc0028ec86e5a6c815f67f6b97d0f92f568295166badd16013100f3b91249a", "0x1b0f12bf16c893648990f3a83f83bd0dc23169ceb37b0996002542c27172ee4b"]],
+ "gb1" :["236eebe2f7022b0a0f16ffcd29c4ad4623ddeca531212565f30b4757f334a80e", "0x2669129239f8138a31193a4db28cf89ca3f4f3a58846f7b19ad320a05d107a4b"],
+ "gb2" :[["0x1eeafc18569d25aa9d455a1d5cdb6197ec209aa3f7d8a1429a65327748ece97d", "0x26640aab2b9b4290965dc3bbca9eb05b3005b59959c817999488e2f5ca9446ea"],
+ ["0x2a681ae342d85857347180307dc30e38e31c76ce8d26b27b41a8e2ccdcbdcebc", "0x17b91219b8a18603d164e6961a9cddd745b8ed2c5570b541ce25ccd1028e2593"]],
+ "z" :[["0x265490dd13abf49ea3442fb5ea0b635e32140e85d60f6c1f12499458c19c09db", "0x1a52eeaddb3cf9ea0914ea157749aca602b60f5ee2da86c4cc6791801cce3e16"],
+ ["0x7f19f8717176e294edb05bad5afed2a27ce6465666334cd6d511b23d0350486", "0x85a13bdd23047a0f863dfc4ce7506262f541d14a7e7e6da25f715097602d7d5"]],
+"IC" :[["23fe51be99cf324124d1050f832459f8f51db95acb745a40de038cfd1690d856", "0x20e9d84953b838cdd8f7dd7119beb20535724d02f2bda1c42503472c6e425704"],["21e866f5696bd1fc6cf268fcce504ae71a23ffbde55130ea652e76828377e540", "0x4c9daf66ed069fff940ef3b0f0f1bd3d01065ca59287292ff544cdf9bd14d02"],["38187aac24fb4fddb3f58dfd4b5c8426f229ff08e1b3691e7e5de2a82fb6d5e", "0x27ac34ac0c88b58cd9009c068b69d371ce6c3d52870e649bad15a12a464c3afa"],["f619f5aee49fbd11251227aa2f50932ff0f180146dc5d6985d3f1e1474a81", "0x2b754e93230102acce70e1f90a9fffa6ebf2914773d5ab0318db576a33b450eb"],["129f735827b54d62487dc8dcfaa70cafbea1692b964f29a6033497dbf7357e67", "0x273cd217a724a0b83b5289c4792a5b95980b783fea16ff54670cb2b7f7f0136e"],["226814ffd18ae1f1e091d630a1186626cf7182fa3e8a897487c4d8302a6df051", "0x1d4df79bc36c8cb92167fbc0e0da1686b2cd86b5c400352b13ac1cd1bec52899"],["1d6f1216d264a36c2330c56b2a137fe95c1a01fdfe34432aa9b28d7da674b53a", "0x18c62d21b612a2835332d9d9e6ceeb2620e9c59795af2e6a69cbc1930b7d76cb"]]}
